// frontend/src/data/interfaces/IBookingState.ts
export interface IBookingState {
  carID: string
  carName: string
  catchAgencyID: string
  catchAgency: string
  catchLocation: string
  catchDate: string
  catchHours: string
  deliveryAgencyId: string
  deliveryAgency: string
  deliveryLocation: string
  deliveryDate: string
  deliveryHours: string
  totalBooking: string
  userID: string
}